Fixed the intermittent level generation bug where corridors couldn't be made
Couple of causes - the minimum room size was only half the width of the leaf, so if
two rooms were generated exactly at half the leaf's side then there's a chance that
there is no overlap. That's easily fixed, just add +1 to the minimum room size.

The other issue was that the room generation died if the room could only be one size,
eg minWidth and maxWidth were the same. Also an easy fix.